Create default PatternCatalogSource if it doesn't exist
It's not possible to bundle a custom resource into an OLM bundle that is
type registry+v1. The controller has to create the default
PatternCataSource object instead.

This code creates a PatternCatalogSource named default in the same
namespace that the controller is running in. (See inline comments about
determining the namespace, it's not straight forward to do.) This has to
happen outside of the manager because once it starts it enters it's
reconcile loop and it's not possible to execute anything else.